% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{IntegerVector}
}
\value{
Integer value - mode of a vector
}
\description{
The function calculates a mode of a given integer vector
}

