% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{perms}
\usage{
perms(x)
}
\arguments{
\item{x}{input vector}
}
\value{
Matrix - the matrix that represents all possible permutations
}
\description{
The function generates all possible combinations of a given vector
}

